LayoutTests/media/modern-media-controls/remaining-time-support/remaining-time-support-expected.txt

Testing the RemainingTimeSupport behavior.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


durationchange event was dispatched
PASS mediaController.controls.timeControl.remainingTimeLabel.value is -media.duration

timeupdate event was dispatched
PASS mediaController.controls.timeControl.remainingTimeLabel.value is media.currentTime - media.duration

PASS successfullyParsed is true

TEST COMPLETE
